[refactor] NumbersGenerator 에서 Integer 대신 Number 교체

+++ src/test/java/lottoTest/LottoServiceTest.java
@@ -32,6 +32,6 @@ public class LottoServiceTest {
         System.out.println(resultLotto.getWinningRate());
         System.out.println(winningAmount / buyCount * 1000);
 
-        Assertions.assertThat(resultLotto.getWinningRate()).isEqualTo(winningAmount / (buyCount * 1000));
+        Assertions.assertThat(resultLotto.getWinningRate()).isEqualTo((double) winningAmount / (buyCount * 1000));
     }
 }
